Refactor track types and improve overall readability
Rearranged the organization of track related types by relocating the 'TrackId' enum to a separate file 'trackId.ts'. This improves the overall readability and maintenance of the project by keeping related types grouped together.

Changes have also been made to improve type safety in 'tracks.ts' through explicit typing. The 'NextTrackRules' interface has been renamed to 'NextTrackRuleSet' to better represent its intended functionality.

Some redundant code in 'requests.ts' was eliminated, removing unnecessary dependencies on 'metaEnv' and simplifying the 'buildHeaders' function.

The visibility checking heuristic in 'TeeTrack.vue' was rollbacked from 'v-if' to 'v-show'

Finally, in 'formTypes.ts', the 'FormField' interface was moved towards the end of the file to enhance readability, and its 'label' property was made optional.
