+++ packages/web/src/types/index.ts
@@ -5,5 +5,6 @@ export * from './naf'
 export * from './otherTypes'
 export * from './programTypes'
 export * from './publicodesObjects'
-export * from './tracksTypes'
+export * from './track/trackType'
 export * from './translationTypes'
+export * from './track/trackId'
