+++ packages/web/src/types/track/trackId.ts
@@ -1,0 +1,26 @@
+export enum TrackId {
+  BuildingProperty = 'track_structure_building_property',
+  BuildingSurface = 'track_structure_building_surface',
+  ContactForm = 'track_form',
+  EnergyReductionPriority = 'track_energy_reduction_priority',
+  EnergyTypes = 'track_energy_types',
+  Goals = 'track_goals',
+  Help = 'track_help',
+  Mobility = 'track_mobility',
+  MobilityEnergy = 'track_mobility_energy',
+  MobilityNumberVehicles = 'track_mobility_number_vehicles',
+  Needs = 'track_needs',
+  Results = 'track_results',
+  Roles = 'track_roles',
+  Sectors = 'track_sectors',
+  Siret = 'track_siret',
+  Status = 'track_status',
+  StrategyAudits = 'track_strategy_audits',
+  StrategyAuditsSelect = 'track_strategy_audits_select',
+  StructureSizes = 'track_structure_sizes',
+  StructureWorkforce = 'track_structure_workforce',
+  WastesMaterials = 'track_wastes_materials',
+  WastesSorting = 'track_wastes_sorting',
+  WastesStake = 'track_wastes_stake',
+  WaterStake = 'track_water_stake',
+}
