+++ packages/web/src/utils/requests.ts
@@ -1,27 +1,9 @@
-import type { FormCallback, MetaEnv } from '@/types/index'
+import type { FormCallback } from '@/types/index'
 // import { toRaw } from 'vue'
 import { remapItem } from './helpers'
 
-export const buildHeaders = (metaEnv: MetaEnv | any, callback: FormCallback) => {
-  // console.log()
-  // console.log('utils > requests > buildHeaders >  metaEnv :', metaEnv)
-  // const mode = 'cors'
-  // console.log('utils > requests > buildHeaders >  mode :', mode)
-
-  // const method = callback.method
-  const headerApiKey = callback.headerApiKey
-  const headers = {...callback.headers}
-  // @ts-ignore
-  const apiKey = metaEnv[callback.envApiKey]
-  // console.log('utils > requests > buildHeaders >  method :', method)
-  // console.log('utils > requests > buildHeaders >  headerApiKey :', headerApiKey)
-  // console.log('utils > requests > buildHeaders >  apiKey :', apiKey)
-
-  // @ts-ignore
-  headers[headerApiKey] = apiKey
-  // console.log('utils > requests > buildHeaders >  headers :', headers)
-
-  return headers
+export const buildHeaders = (callback: FormCallback) => {
+  return { ...callback.headers }
 }
 
 export const sendApiRequest = async (
@@ -36,11 +18,10 @@ export const sendApiRequest = async (
   // console.log('utils > requests > sendApiRequest >  usedTrack :', usedTrack)
   // console.log('utils > requests > sendApiRequest >  props :', props)
 
-  const metaEnv = import.meta.env
   // console.log('utils > requests > sendApiRequest >  metaEnv :', metaEnv)
   const url = callback.url
   const method = callback.method
-  const headers = buildHeaders(metaEnv, callback)
+  const headers = buildHeaders(callback)
   // console.log('utils > requests > sendApiRequest >  url :', url)
   // console.log('utils > requests > sendApiRequest >  method :', method)
   // console.log('utils > requests > sendApiRequest >  headers :', headers)
